% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_card.R
\name{create_card}
\alias{create_card}
\title{create bslib card}
\usage{
create_card(
  row_prepped,
  type_card = c("day", "year"),
  type_fz = c("Velos", "Autos")
)
}
\arguments{
\item{row_prepped}{single row of data prepared for the card}

\item{type_card}{type of card, "day" or "year"}

\item{type_fz}{string, used in label, "Velos" or "Autos"}
}
\value{
bslib card
}
\description{
create card for showing the number of vehicles in current year or on latest day
}
