% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verkehrszaehlungen-package.R
\docType{package}
\name{verkehrszaehlungen-package}
\alias{verkehrszaehlungen}
\alias{verkehrszaehlungen-package}
\title{verkehrszaehlungen: Count of cars/bikes at selected stations in the city of Zurich}
\description{
Showing aggregate data, variant for cars or bikes.
}
\author{
\strong{Maintainer}: Sara Hänzi \email{sara.haenzi@zuerich.ch}

Other contributors:
\itemize{
  \item Statistik Stadt Zürich \email{statistik@zuerich.ch} [copyright holder, funder]
}

}
\keyword{internal}
